pyproject.toml
[tool.poetry]
name = "nonebot-plugin-tetrio-stats"
version = "0.2.0"
description = "一个在nonebot2中查询TETRIO玩家数据的插件，同时支持绑定TETRIO账号"
authors = ["scdhh <wallfjjd@gmail.com>"]
packages = [{ include = 'nonebot-plugin-tetrio-stats' }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
asyncio = "^3.4.3"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot2 = "^2.0.0-beta.2"
repository = "https://github.com/shoucandanghehe/nonebot-plugin-tetrio-stats"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
